Tidy up the backend
I have placed all the CSS, JS, HTML, and backend related files in the app folder for easy management.
