+++ CITS3403-Project/.gitignore
@@ -1,4 +1,3 @@
 /venv
 /instance
 /__pycache__
-/app
