+++ CITS3403-Project/app/static/js/hide_password.js
